backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
gunicorn==23.0.0
tensorflow-cpu==2.16.2
